+++ src/test/java/com/gdschongik/gdsc/domain/email/domain/HongikUnivEmailValidatorTest.java
@@ -7,7 +7,6 @@ import com.gdschongik.gdsc.domain.member.domain.Member;
 import com.gdschongik.gdsc.global.common.constant.MemberConstant;
 import com.gdschongik.gdsc.global.exception.CustomException;
 import com.gdschongik.gdsc.global.exception.ErrorCode;
-import com.gdschongik.gdsc.global.util.email.HongikUnivEmailValidator;
 import java.util.Optional;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
